requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.0.3
ecdsa==0.18.0
Flask==2.2.2
flask-cognito-auth==1.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.3
idna==3.4
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.1.1
mod-wsgi==4.9.4
pyasn1==0.4.8
python-dotenv==0.21.0
python-jose==3.3.0
requests==2.28.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.41
urllib3==1.26.12
Werkzeug==2.2.2
wget==3.2
WTForms==3.0.1
